File Explosion
Exploded ytmp3 into multiple files
