+++ src/lib/ytmp3/index.ts
@@ -1,0 +1,8 @@
+export const UA = { "User-Agent": "SP2YT/1.0 (your-email@example.com)" };
+export const runtime = "nodejs";
+export const dynamic = "force-dynamic";
+
+export * from "./info";
+export * from "./parse";
+export * from "./process";
+export * from "./helpers";
