+++ src/lib/ytmp3/types.ts
@@ -1,0 +1,5 @@
+export type ReqBody = {
+  urls: string[];
+  album?: string; // optional album override for the whole batch
+  concurrency?: number; // 1–3 recommended
+};
